Move away from private file names
