+++ src/fast_bioservices/bigg/bigg.py
@@ -1,8 +1,9 @@
 import json
 from typing import Any, Literal, Mapping, Optional
 
-from fast_bioservices._base import BaseModel
-from fast_bioservices._fast_http import FastHTTP
+from fast_bioservices.base import BaseModel
+from fast_bioservices.fast_http import FastHTTP
+from fast_bioservices.settings import default_workers
 
 
 class BiGG(BaseModel, FastHTTP):
