+++ src/fast_bioservices/settings.py
@@ -10,3 +10,4 @@ log_filepath: Path = Path(_root_cache_dir, "fast_bioservices.log")
 _root_cache_dir.mkdir(parents=True, exist_ok=True)
 cache_dir.mkdir(parents=True, exist_ok=True)
 log_filepath.touch(exist_ok=True)
+default_workers: int = 4
